pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "h5py>=3.16.0",
    "keras>=3.13.2",
    "numpy>=2.4.3",
    "pillow>=12.1.1",
    "tensorflow-cpu>=2.20.0",
]
